refactor: LoginController
- Config와의 의존관계 제거
